+++ juliapreview2.c
@@ -113,7 +113,7 @@ int main ()
     draw_mandelbrot(mandelbrot_screen, mandelbrot_region, render_rect,
 		    colormap, MAXITERS);
     Uint32 stop = SDL_GetTicks();
-    fprintf(stderr, "  Mandelbrot took %lums\n", stop - start);
+    fprintf(stderr, "  Mandelbrot took %lums\n", (long unsigned) stop - start);
   }
   // Assign a c and render an initial Julia
   complex c = {.233, .53780};
